Completed implementation for LoanServiceImpl

+++ src/main/java/com/vdavid/apps/librarymanagementsystem/model/Item.java
@@ -1,4 +1,4 @@
 package com.vdavid.apps.librarymanagementsystem.model;
 
-public record Item(Integer uniqueId, Integer ItemId, ItemType itemType, String title) {
+public record Item(Long uniqueId, Integer ItemId, ItemType itemType, String title) {
 }
